Logically organised into files
